refactor: reorganize directories and imports
