--- server/utils/dayUtils.ts
@@ -1,11 +1,0 @@
-export const weekdays: string[] = [
-	"sunday",
-	"monday",
-	"tuesday",
-	"wednesday",
-	"thursday",
-	"friday",
-	"saturday",
-];
-
-export const getWeekday = (day: number) => weekdays[day];
