+++ server/utils/getWeekday.ts
@@ -1,0 +1,13 @@
+const weekdays: string[] = [
+	"sunday",
+	"monday",
+	"tuesday",
+	"wednesday",
+	"thursday",
+	"friday",
+	"saturday",
+];
+
+const getWeekday = (day: number) => weekdays[day];
+
+export default getWeekday;
