+++ server/utils/shouldIDeploy.ts
@@ -1,1 +1,3 @@
-export const shouldIDeploy = (day: number) => day > 0 && day < 5;
+const shouldIDeploy = (day: number) => day > 0 && day < 5;
+
+export default shouldIDeploy;
